+++ src/main/java/nl/tudelft/skills/service/GitLabReleaseClient.java
@@ -25,10 +25,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import nl.tudelft.skills.dto.old.view.ReleaseDTO;
+import nl.tudelft.skills.dto.view.ReleaseDTO;
 
 @Service
-public class GitLabClient {
+public class GitLabReleaseClient {
 	private final WebClient webClient = WebClient.builder().baseUrl("https://gitlab.ewi.tudelft.nl/").build();
 
 	/**