+++ CHANGELOG.md
@@ -12,6 +12,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## Added
+
+## Changed
+
+## Fixed
+
+
+## [2526.0.0]
+
+## Added
+- [Everyone] A changelog is displayed to users after each new release. (@bbakos)
+
+## Changed
+- [Everyone] The UI
+- [Editor] On edit, the minTasks value of a choice task is clipped to be inside the correct bounds. (@ffiedrich)
+- [Everyone] The action indicator is changed to in the "glass" style. (@rwbackx)
+- [Student] The cycle warning is not shown to students anymore. (@ffiedrich)
+- [Student] The information about a checkpoint having passed is changed to a dialog. (@ffiedrich)
+
+## Fixed
+- [Editor] Fixed moving subtasks outside choice tasks. (@ffiedrich)
+
+
+## [2.2.8]
+
+## Added
 - [Everyone] Functionality to add and complete choice tasks. Choice tasks contain a set of regular tasks, of which at least k need to be completed.
 
 ## Changed